fix linter errors by refactoring class APp
