+++ app.rb
@@ -6,9 +6,14 @@ require './student'
 require './teacher'
 require './classroom'
 require './data/load'
+require './modules/list_records'
+require './modules/create_records'
 require 'json'
 
 class App
+  include ListItem
+  include CreateItem
+
   def initialize
     @load = Load.new
     @books = []
@@ -27,129 +32,6 @@ class App
       end
       option_selector(option)
     end
-  end
-
-  def list_books
-    @load.load_books
-    if @load.books.empty?
-      puts 'There are no books in the library'
-    else
-      @load.books.each do |book|
-        puts "Title: #{book.title}, Author: #{book.author}".capitalize
-      end
-    end
-  end
-
-  def list_people
-    @load.load_people
-    if @load.people.empty?
-      puts 'There are no people in the library'
-    else
-      @load.people.each do |person|
-        if person.is_a?(Teacher) && person.specialization
-          puts "[Teacher] ID: #{person.id}, Name: #{person.name}, Age: #{person.age}, Specialization: #{person.specialization}"
-        else
-          puts "[Student] ID: #{person.id}, Name: #{person.name}, Age: #{person.age}"
-        end
-      end
-    end
-  end
-
-  def list_rentals_person
-    puts 'Select a person from the following list by number (not id)'
-    @people.each_with_index { |person, index| puts "#{index}) #{person.name}" }
-    person_index = gets.chomp.to_i
-    show_rentals_person(person_index)
-  end
-
-  def show_rentals_person(person_index)
-    person = @people[person_index]
-    puts "#{person.name} has rented the following books:"
-    person.rentals.each do |rental|
-      puts "Date: #{rental.date}, Book: #{rental.book.title} by #{rental.book.author}"
-    end
-  end
-
-  def create_person
-    puts 'Do you want to create a student (1) or a teacher (2)? [Input the number]'
-    option = gets.chomp
-
-    case option
-    when '1'
-      create_student
-    when '2'
-      create_teacher
-    else
-      puts 'That is not a valid input'
-    end
-  end
-
-  def create_student
-    id = @people.length + 1
-    puts 'Name:'
-    name = gets.chomp
-    puts 'Age:'
-    age = gets.chomp.to_i
-    puts 'Has parent permission? [Y/N]?'
-    permission = gets.chomp.downcase
-    parent_permission = permission == 'y'
-    student = Student.new(id, age, name, parent_permission)
-    add_person(student)
-    save_person(student)
-    puts 'Student created successfully'
-  end
-
-  def create_teacher
-    puts 'Name:'
-    name = gets.chomp
-    puts 'Age:'
-    age = gets.chomp.to_i
-    puts 'Specialization:'
-    specialization = gets.chomp
-    teacher = Teacher.new(name, age, specialization)
-    add_person(teacher)
-    save_person(teacher)
-    puts 'Teacher created successfully'
-  end
-
-  def create_book
-    puts 'Title:'
-    title = gets.chomp
-    puts 'Author:'
-    author = gets.chomp
-    book = Book.new(title, author)
-    add_book(book)
-    save_book(book)
-    puts 'Book created successfully'
-  end
-
-  def create_rental
-    puts 'Select a book from the following list by number'
-    @books.each_with_index { |book, index| puts "#{index}) #{book.title}" }
-    book_index = gets.chomp.to_i
-
-    if book_index.negative? || book_index >= @books.length
-      puts 'Invalid book selection'
-      return
-    end
-
-    puts 'Select a person from the following list by number (not id)'
-    @people.each_with_index { |person, index| puts "#{index}) #{person.name}" }
-    person_index = gets.chomp.to_i
-
-    if person_index.negative? || person_index >= @people.length
-      puts 'Invalid person selection'
-      return
-    end
-
-    puts 'Date:'
-    date = gets.chomp
-
-    book = @books[book_index]
-    puts "Selected book: #{book.title} by #{book.author}"
-
-    rental = Rental.new(date, book, @people[person_index])
-    add_rental(rental)
   end
 
   private
@@ -198,18 +80,18 @@ class App
 
   def save_person(person)
     people_data = []
-  
+
     if File.exist?('./data/people.json')
       file = File.read('./data/people.json')
       people_data = JSON.parse(file) unless file.empty?
     end
-  
-    if person.is_a?(Teacher)
-      people_data << { id: person.id, name: person.name, age: person.age, specialization: person.specialization }
-    else
-      people_data << { id: person.id, name: person.name, age: person.age }
-    end
-  
+
+    people_data << if person.is_a?(Teacher)
+                     { id: person.id, name: person.name, age: person.age, specialization: person.specialization }
+                   else
+                     { id: person.id, name: person.name, age: person.age }
+                   end
+
     File.write('./data/people.json', JSON.generate(people_data))
   end
